ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
textblob==0.17.1
vaderSentiment==3.3.2
sentence-transformers==2.2.2
datasets==2.14.6
nltk==3.8.1
spacy==3.7.2
googlesearch-python==1.2.3
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78

# Additional dependencies for FakeNewsNet training
xgboost==2.0.0
joblib==1.3.2
pathlib2==2.3.7

# Enhanced AI Dependencies  
playwright==1.40.0
aiohttp==3.9.1
feedparser==6.0.10